refactor. 옵션 컴포넌트 관심사 분리 #40
